Refactor imports and clean up code in avarias_a_receber app
