+++ avarias_a_receber/apps.py
@@ -1,6 +1,5 @@
 from django.apps import AppConfig
 
-
 class AvariasAReceberConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'avarias_a_receber'
