+++ avarias_a_receber/models.py
@@ -1,6 +1,5 @@
 from django.db import models
 
-# Create your models here.
 class Avaria(models.Model):
     localizacao = models.CharField(max_length=255)
     instalacao = models.CharField(max_length=255)
