+++ avarias_a_receber/serializers.py
@@ -1,12 +1,10 @@
 from rest_framework import serializers
-from .models import Avaria
-
+from .models import Avaria, Material  # Corrected import
 
 class AvariaSerializer(serializers.ModelSerializer):
     class Meta:
         model = Avaria
         fields = '__all__'
-
 
 class MaterialSerializer(serializers.ModelSerializer):
     class Meta:
