+++ avarias_a_receber/urls.py
@@ -1,9 +1,9 @@
 from django.urls import path
-from ..Avarias_a_receber import views # type: ignore
+from . import views  # Corrected import
 
 urlpatterns = [
-    path('avarias/', views.AvariaListCreate.as_view(), name='avarias-list-create'),
-    path('avarias/<int:pk>/', views.AvariaRetrieveUpdateDestroy.as_view(), name='avarias-retrieve-update-destroy'),
+    path('', views.AvariaListCreate.as_view(), name='avarias-list-create'),
+    path('<int:pk>/', views.AvariaRetrieveUpdateDestroy.as_view(), name='avarias-retrieve-update-destroy'),
     path('materiais/', views.MaterialListCreate.as_view(), name='materiais-list-create'),
     path('materiais/<int:pk>/', views.MaterialRetrieveUpdateDestroy.as_view(), name='materiais-retrieve-update-destroy'),
 ]
